requirements.txt
faker==4.14.2
PyYAML==5.3.1
kafka-python==2.0.2
elasticsearch==7.10.0